feat: #110: use server components to render the explore page
