--- app/api/stats/route.ts
@@ -1,1 +1,0 @@
-export { GET } from '@/shared/api/server/stats';
